+++ src/app/products/product/click-prevent.directive.ts
@@ -1,0 +1,15 @@
+import {Directive, EventEmitter, HostListener, Output} from '@angular/core';
+
+@Directive({
+    selector: '[appClickPrevent]'
+})
+
+export class ClickPreventDirective {
+    @Output() appClickPrevent = new EventEmitter();
+
+    @HostListener('click', ['$event'])
+    onClick(event) {
+        event.preventDefault();
+        this.appClickPrevent.emit(event);
+    }
+}