+++ src/app/products/products.module.ts
@@ -4,7 +4,8 @@ import {CommonModule} from '@angular/common';
 import {ListComponent} from './list/list.component';
 import {ProductComponent} from './product/product.component';
 
-import {ListService} from './list.service';
+import {ListService} from './products.service';
+import {ClickPreventDirective} from './product/click-prevent.directive';
 
 @NgModule({
     imports: [
@@ -12,7 +13,8 @@ import {ListService} from './list.service';
     ],
     declarations: [
         ListComponent,
-        ProductComponent
+        ProductComponent,
+        ClickPreventDirective
     ],
     providers: [ListService],
     exports: [ListComponent]