+++ CHANGELOG.md
@@ -11,3 +11,9 @@
 - Add cart service, which returns stub list of items (not connected to real product list yet)
 - Render stub items from cart service in cart component
 - Render alternative block, if cart list is empty
+
+####[21.08.2018] Task2
+- Add modules for Products and Cart components, register modules in App component
+- Add child Product component
+- Realize parent-child communication withing Product List component
+- Add link 'Show/Hide' to toggle product description visibility. Apply directive to prevent default link click behaviour.